Move scrape route to main app and add session handler
